+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Improved the performance when querying transactions. ([#3464](https://github.com/craftcms/commerce/issues/3464))
 - Fixed a bug where the footer delete action wasn’t deleting shipping methods. ([#3444](https://github.com/craftcms/commerce/issues/3444))
 
 ## 4.5.4 - 2024-04-24